script upgrade

+++ scripts/RunAnalysis-mc.sh
@@ -6,18 +6,23 @@ for X in `seq 1 1 100`; do
 	if [ -e ${output}.root ];then
 	    echo "...was already done."
 	else
-            scriptname="mc$1$X.sh"
-            if [ -e ${scriptname} ]; then
-        	echo "...already in process."
-    	    else
-		echo "...PROCESSING MC $1..."
-                echo "#!/bin/bash" >> ${scriptname}
-                echo "./rawanalysis MC$1 -local -mode mc -fin file:${input} -n ${output} -abort" >> ${scriptname}
-                echo >> ${scriptname}
-                echo "rm -f $PWD/${scriptname}" >> ${scriptname}
-                chmod u+x ${scriptname}
-                ./${scriptname} &> ${output}.log
-		echo "...done."
+	    wmc_script="run_wmc-$1-$X.sh"
+	    if [ -e ${wmc_script} ]; then
+		echo "... WMC is stil processing."
+	    else
+	            scriptname="mc$1$X.sh"
+        	    if [ -e ${scriptname} ]; then
+	        	echo "...already in process."
+	    	    else
+			echo "...PROCESSING MC $1..."
+	                echo "#!/bin/bash" >> ${scriptname}
+	                echo "./rawanalysis MC$1 -local -mode mc -fin file:${input} -n ${output} -abort" >> ${scriptname}
+	                echo >> ${scriptname}
+	                echo "rm -f $PWD/${scriptname}" >> ${scriptname}
+	                chmod u+x ${scriptname}
+	                ./${scriptname} &> ${output}.log
+			echo "...done."
+		    fi
             fi
         fi
     fi
